+++ packages/create/test/unit/cli/add/annotations-to-odata.test.ts
@@ -1,6 +1,7 @@
 import type { ManifestNamespace } from '@sap-ux/project-access';
-import type { Editor } from 'mem-fs-editor';
+import type { Store } from 'mem-fs';
 import type { ToolsLogger } from '@sap-ux/logger';
+import type { Editor, create } from 'mem-fs-editor';
 import { Command } from 'commander';
 import { addAnnotationsToOdataCommand } from '../../../../src/cli/add/annotations-to-odata';
 import * as tracer from '../../../../src/tracing/trace';
@@ -11,8 +12,8 @@ import * as adp from '@sap-ux/adp-tooling';
 import * as projectAccess from '@sap-ux/project-access';
 import { readFileSync } from 'fs';
 import { join } from 'path';
+import * as oDataWriter from '@sap-ux/odata-service-writer';
 
-const appManifest = readFileSync(join(__dirname, '../../../fixtures/adaptation-project', 'manifest.json'), 'utf-8');
 const descriptorVariant = JSON.parse(
     readFileSync(join(__dirname, '../../../fixtures/adaptation-project', 'manifest.appdescr_variant'), 'utf-8')
 );
@@ -20,19 +21,13 @@ const descriptorVariant = JSON.parse(
 jest.mock('prompts');
 jest.mock('@sap-ux/adp-tooling');
 
-const mockAppInfo = { ExampleApp: { manifestUrl: 'https://sap.example' } };
-const abapServicesMock = {
-    getAppInfo: jest.fn().mockResolvedValue(mockAppInfo),
-    getManifest: jest.fn().mockResolvedValue(JSON.parse(appManifest))
-};
-
 const mockDataSources = {
     'annotation': {
         'settings': { 'localUri': 'localService/annotation.xml' },
         'type': 'ODataAnnotation',
         'uri': "/path/to/annotation;v=2/Annotations(TechnicalName='annotation',Version='0001')/$value/?sap-language=EN"
     },
-    'service': {
+    'mainService': {
         'settings': {
             'annotations': ['annotation'],
             'localUri': 'localService/mockdata/metadata.xml'
@@ -42,39 +37,19 @@ const mockDataSources = {
     }
 } as unknown as Record<string, ManifestNamespace.DataSource>;
 
-jest.mock('@sap-ux/system-access', () => {
-    return {
-        ...jest.requireActual('@sap-ux/system-access'),
-        createAbapServiceProvider: () => {
-            return {
-                getAppIndex: jest.fn().mockReturnValue({
-                    getAppInfo: abapServicesMock.getAppInfo
-                }),
-                getLayeredRepository: jest.fn().mockReturnValue({
-                    getManifest: abapServicesMock.getManifest
-                })
-            };
-        }
-    };
-});
-
 describe('add/annotations', () => {
     let loggerMock: ToolsLogger;
     const memFsEditorMock = {
-        create: jest.fn().mockReturnValue({
-            commit: jest.fn().mockImplementation((cb) => cb())
-        })
+        commit: jest.fn().mockImplementation((cb) => cb())
     };
     const traceSpy = jest.spyOn(tracer, 'traceChanges');
     const generateChangeSpy = jest
         .spyOn(adp, 'generateChange')
         .mockResolvedValue(memFsEditorMock as Partial<Editor> as Editor);
     const getArgv = (...arg: string[]) => ['', '', 'annotations', ...arg];
-    const mockAnswers = {
-        targetODataSource: 'mainService',
-        targetODataUrl: '/sap/opu/odata/test',
-        annotationUri: '/sap/opu/odata/test/annotation',
-        maxAge: 60
+    const mockAnswers: adp.AddAnnotationsAnswers = {
+        id: 'mainService',
+        fileSelectOption: 2
     };
     const promptYUIQuestionsSpy = jest.spyOn(common, 'promptYUIQuestions').mockResolvedValue(mockAnswers);
     jest.spyOn(adp, 'getAdpConfig').mockResolvedValue({
@@ -84,8 +59,21 @@ describe('add/annotations', () => {
         }
     });
     jest.spyOn(validations, 'validateAdpProject').mockResolvedValue(undefined);
-    jest.spyOn(adp, 'getManifestDataSources').mockResolvedValue(mockDataSources);
     jest.spyOn(adp, 'getPromptsForAddAnnotationsToOData').mockImplementation(() => []);
+    jest.spyOn(adp.ManifestService, 'initMergedManifest').mockResolvedValue({
+        fetchBaseManifest: jest.fn(),
+        fetchMergedManifest: jest.fn(),
+        getManifest: jest.fn(),
+        fetchAppInfo: jest.fn(),
+        getManifestDataSources: jest.fn().mockReturnValue(mockDataSources),
+        getDataSourceMetadata: jest.fn().mockResolvedValue('<>metadata</>')
+    } as unknown as adp.ManifestService);
+    jest.spyOn(oDataWriter, 'getAnnotationNamespaces').mockReturnValue([
+        {
+            alias: 'alias',
+            namespace: 'namespace'
+        }
+    ]);
 
     const appRoot = join(__dirname, '../../../fixtures');
 
@@ -159,11 +147,40 @@ describe('add/annotations', () => {
         expect(traceSpy).toBeCalled();
     });
 
+    test('should not fetch metadata when file path is provided', async () => {
+        jest.spyOn(oDataWriter, 'getAnnotationNamespaces');
+        mockAnswers.fileSelectOption = 1;
+        mockAnswers.filePath = 'path/to/file.xml';
+        const command = new Command('annotations');
+        addAnnotationsToOdataCommand(command);
+        await command.parseAsync(getArgv(appRoot, '--simulate'));
+        mockAnswers.fileSelectOption = 2;
+        mockAnswers.filePath = undefined;
+
+        expect(promptYUIQuestionsSpy).toBeCalled();
+        expect(generateChangeSpy).toBeCalled();
+        expect(traceSpy).toBeCalled();
+        expect(oDataWriter.getAnnotationNamespaces).not.toBeCalled();
+    });
+
     test('should fail with authentication error after 3 attempts', async () => {
-        jest.spyOn(adp, 'getManifestDataSources').mockRejectedValueOnce({
-            message: '401:Unauthorized',
-            response: { status: 401 }
-        });
+        jest.spyOn(adp.ManifestService, 'initMergedManifest')
+            .mockRejectedValueOnce({
+                message: '401:Unauthorized',
+                response: { status: 401 }
+            } as unknown as adp.ManifestService)
+            .mockRejectedValueOnce({
+                message: '401:Unauthorized',
+                response: { status: 401 }
+            } as unknown as adp.ManifestService)
+            .mockRejectedValueOnce({
+                message: '401:Unauthorized',
+                response: { status: 401 }
+            } as unknown as adp.ManifestService)
+            .mockRejectedValueOnce({
+                message: '401:Unauthorized',
+                response: { status: 401 }
+            } as unknown as adp.ManifestService);
 
         const command = new Command('annotations');
         addAnnotationsToOdataCommand(command);
@@ -174,14 +191,16 @@ describe('add/annotations', () => {
             'Authentication failed. Please check your credentials. Login attempts left: 2'
         );
         expect(loggerMock.debug).not.toBeCalledWith();
-        expect(promptYUIQuestionsSpy).toBeCalled();
-        expect(generateChangeSpy).toBeCalled();
+        expect(promptYUIQuestionsSpy).not.toBeCalled();
+        expect(generateChangeSpy).not.toBeCalled();
     });
 
     test('should fail when no data sources found in base application manifest', async () => {
-        jest.spyOn(adp, 'getManifestDataSources').mockRejectedValueOnce(
-            new Error('No data sources found in the manifest')
-        );
+        jest.spyOn(adp.ManifestService, 'initMergedManifest').mockResolvedValueOnce({
+            getManifestDataSources: jest.fn().mockImplementation(() => {
+                throw new Error('No data sources found in the manifest');
+            })
+        } as unknown as adp.ManifestService);
 
         const command = new Command('annotations');
         addAnnotationsToOdataCommand(command);